+++ src/test/testdata/expected/lsl/generics.lsl
@@ -51,10 +51,13 @@ automaton A : HashMap<K, V> {
         _genericFunWithParams<Int, Int, Int>(4, 5);
     }
 
-    fun *.genericTypeDefBlockReturnType(): HashMap<Int, Int> {
+    fun *.genericTypeDefBlockReturnType(x: K): HashMap<Int, Int> {
         var newHashMap: HashMap<Int, Int> = new A<Int, Int>(state = Initialized);
         A(newHashMap)._genericProc<Int, Int, Int>();
         result = newHashMap;
         action PLAIN_GENERIC_ACTION<Int>(5, 6);
+        if (x is HashMap<Int, Int>) {
+
+        }
     }
 }
