+++ src/test/testdata/lsl/generics.lsl
@@ -79,10 +79,13 @@ automaton A
         _genericFunWithParams<Int, Int, Int>(4, 5);
     }
 
-    fun *.genericTypeDefBlockReturnType (): HashMap <Int, Int> {
+    fun *.genericTypeDefBlockReturnType (x: K): HashMap <Int, Int> {
         var newHashMap: HashMap<Int, Int> = new A<Int, Int>(state = Initialized);
         A(newHashMap)._genericProc<Int, Int, Int>();
         result = newHashMap;
         action PLAIN_GENERIC_ACTION<Int>(5, 6);
+        if (x is HashMap <Int, Int>) {
+
+        }
     }
 }
